adds delete dialog for each employee

+++ src/app/modules/planning/components/employee-list/employee-list.component.ts
@@ -1,5 +1,7 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
+import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
 import { faTrashAlt } from '@fortawesome/free-solid-svg-icons';
+import { AlertComponent } from '@shared/dialogs/alert/alert.component';
 import { PeriodService } from '@shared/services/period.service';
 
 @Component({
@@ -18,15 +20,32 @@ export class EmployeeListComponent implements OnInit {
   maxHours: number = 48;
   faTrashAlt = faTrashAlt;
 
-  constructor(private periodService: PeriodService) { }
+  constructor(private periodService: PeriodService, private dialog: MatDialog) {}
 
   ngOnInit(): void {
   }
 
   deleteEmployee(): void{
-    this.periodService.deleteEmployee(this.periodId, this.employee._id).subscribe((res) => {
-      console.log(res);
-      this.deleteEmployeeEvent.emit();
-    })
+
+    const dialogConfig = new MatDialogConfig();
+    dialogConfig.disableClose = true;
+    dialogConfig.data = { 
+      title: `Quitar empleado ${this.employee.lastName} ${this.employee.firstName}`,
+      text: `Desea quitar al empleado ${this.employee.lastName} ${this.employee.firstName} de la planificación?`,
+      btnClass: "btn-danger"
+     };
+
+    this.dialog.open(AlertComponent, dialogConfig)
+    .afterClosed()
+    .subscribe((confirm: boolean)  => {
+      if(confirm){
+        
+        this.periodService.deleteEmployee(this.periodId, this.employee._id).subscribe((res) => {
+          console.log(res);
+          this.deleteEmployeeEvent.emit();
+        });
+
+      }
+    });
   }
 }
